% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans.R
\name{as.emm}
\alias{as.emm}
\alias{as.list.emm}
\title{Convert to and from \code{emm} objects}
\usage{
as.emm(object, ...)

\method{as.list}{emm}(x, ...)
}
\arguments{
\item{object}{Object to be converted to class \code{emm}. It may
be a \code{list} returned by \code{as.list.emm}, or a \code{ref.grid}
or \code{lsmobj} object created by \pkg{emmeans}'s predecessor, the 
\pkg{lsmeans} package. An error is thrown if \code{object} cannot
be converted.}

\item{...}{In \code{as.emm}, additional arguments passed to 
\code{\link{update.emm}} before returning the object. This
argument is ignored in \code{as.list.emm}}

\item{x}{An \code{emm} object}
}
\value{
\code{as.emm} returns an object of class \code{emm}.

\code{as.list.emm} returns an object of class \code{list}.
}
\description{
These are useful utility functions for creating a compact version of an
\code{emm} object that may be saved and later reconstructed, or for
converting old \code{ref.grid} or \code{lsmobj} objects into \code{emm}
objects.
}
\details{
An \code{emm} object is an S4 object, and as such cannot be saved in a
text format or saved without a lot of overhead. By using \code{as.list},
the essential parts of the object are converted to a list format that can be
easily and compactly saved for use, say, in another session or by another user.
Providing this list as the arguments for \code{\link{emmobj}} allows the user 
to restore a working \code{emm} object.
}
\examples{
pigs.lm <- lm(log(conc) ~ source + factor(percent), data = pigs)
pigs.sav <- as.list(ref_grid(pigs.lm))

pigs.anew <- as.emm(pigs.sav)
emmeans(pigs.anew, "source")

\dontrun{
## Convert an entire workspace saved from an old **lsmeans** session
remove(".Last.ref.grid")
for (nm in names(.GlobalEnv)) {
    obj <- get(nm)
    if (is(obj, "ref.grid"))
        assign(nm, as.emm(obj))
}
unloadNamespace("lsmeans")
## Now, save your updated workspace
}
}
\seealso{
\code{\link{emmobj}}
}
