% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans.R
\name{as.list.emm}
\alias{as.list.emm}
\title{Coerce an \code{emm} object to a list}
\usage{
\method{as.list}{emm}(x, ...)
}
\arguments{
\item{x}{An \code{emm} object}

\item{...}{(Required argument, but ignored)}
}
\description{
This is a useful utility for creating a compact version of an \code{emm} object
that may be saved and later reconstructed.
}
\details{
An \code{emm} object is an S4 object, and as such cannot be saved in a
text format or saved without a lot of overhead. By using \code{as.list},
the essential parts of the object are converted to a list format that can be
easily and compactly saved for use, say, in another session or by another user.
Providing this list as the arguments for \code{\link{emmobj}} allows the user 
to restore a working \code{emm} object. See the example below.
}
\examples{
pigs.lm <- lm(log(conc) ~ source + factor(percent), data = pigs)
pigs.sav <- as.list(ref_grid(pigs.lm))

pigs.anew <- do.call(emmobj, pigs.sav)
emmeans(pigs.anew, "source")
}
\seealso{
\code{\link{emmobj}}
}
