% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emm-methods.R
\name{str.emm}
\alias{str.emm}
\alias{print.emm}
\alias{vcov.emm}
\title{Miscellaneous methods for \code{emm} objects}
\usage{
\method{str}{emm}(object, ...)

\method{print}{emm}(x, ...)

\method{vcov}{emm}(object, ...)
}
\arguments{
\item{object}{An \code{emm} object}

\item{...}{(required but not used)}

\item{x}{An \code{emm} object}
}
\value{
The \code{vcov} method returns a ymmetric matrix of variances and
  covariances for \code{predict.emm(object, type = "lp")}
}
\description{
Miscellaneous methods for \code{emm} objects
}
