% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{models}
\title{Models supported in \pkg{emmeans}}
\description{
Documentation for models has been moved to a vignette. To access it,
use \href{../doc/models.html}{\code{vignette("models", "emmeans")}}.
}
