% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast}
\alias{contrast}
\alias{contrast.emm}
\alias{pairs.emm}
\alias{coef.emm}
\title{Contrasts and linear functions of EMMs}
\usage{
contrast(object, ...)

\method{contrast}{emm}(object, method = "eff", interaction = FALSE, by,
  offset = NULL, name = "contrast", options = get_emm_option("contrast"),
  log.contrast = get_emm_option("log.contrast"), adjust, ...)

\method{pairs}{emm}(x, reverse = FALSE, ...)

\method{coef}{emm}(object, ...)
}
\arguments{
\item{object}{An object of class \code{emm}}

\item{...}{Additional arguments passed to other methods}

\item{method}{Character value giving the root name of a contrast method (e.g.
\code{"pairwise"} -- see \link{emmc-functions}). Alternatively, a named 
\code{list} of coefficients (for a contrast or linear function) that must
each conform to the number of results in each \code{by} group.
In a multi-factor situation, the factor levels are combined and treated
like a single factor.}

\item{interaction}{Character vector or logical value. If this is specified,
\code{method} is ignored. See the \dQuote{Interaction contrasts} section
below for details.}

\item{by}{Character names of variable(s) to be used for ``by'' groups. The
contrasts or joint tests will be evaluated separately for each combination
of these variables. If \code{object} was created with by groups, those are
used unless overridden. Use \code{by = NULL} to use no by groups at all.}

\item{offset}{Numeric vector of the same length as each \code{by} group.
These values are added to their respective linear estimates. (It is ignored
when \code{interaction} is specified.)}

\item{name}{Character name to use to override the default label for contrasts
used in table headings or subsequent contrasts of the returned object.}

\item{options}{If non-\code{NULL}, a named \code{list} of arguments to pass
to \code{\link{update.emm}}, just after the object is constructed.}

\item{log.contrast}{Character value matching \code{"ratio"} or 
\code{"difference"}. This only has an effect when a log or logit 
transformation is in force, \emph{and} every set of contrast coefficients 
sums to zero (true contrasts). In such a case, contrasts can (usually) be 
interpreted as logs of ratios. Subsequently if summarizing using \code{type
= "response"}, do we want those results presented as ratios (specify 
\code{"ratio"}), or as differences of back-transformed EMMs (specify 
\code{"difference"})?}

\item{adjust}{Method to use for adjusting \emph{P} values and confidence
intervals. This is an alternative to specifying it as part of
\code{options}.}

\item{x}{An \code{emm} object}

\item{reverse}{Logical value - determines whether to use \code{"pairwise"} (if \code{TRUE}) or \code{"revpairwise"} (if \code{FALSE}).}
}
\value{
\code{contrast} and \code{pairs} return an object of class \code{emm}. Its grid will correspond to the levels of the contrasts and any \code{by} variables.

\code{coef} returns a \code{data.frame} containing the object's grid, along with columns named \code{c.1, c.2, ...} containing the contrast coefficients. If
}
\description{
These methods provide for follow-up analyses of \code{emm} objects:
Contrasts, pairwise comparisons, tests, and confidence intervals. They may
also be used to compute arbitrary linear functions of predictions or EMMs.
}
\note{
When \code{object} has a nesting structure (this can be seen via
  \code{str(object)}), then any grouping factors involved are forced into
  service as \code{by} variables, and the contrasts are thus computed
  separately in each nest. This in turn may lead to an irregular grid in the
  returned \code{emm} object, which may not be valid for subsequent
  \code{emmeans} calls.
}
\section{Pairs method}{

The call \code{pairs(object)} is equivalent to \code{contrast(object, method = "pairwise")}; and \code{pairs(object, reverse = TRUE)} is the same as \code{contrast(object, method = "revpairwise")}.
}

\section{Interaction contrasts}{

When \code{interaction} is specified, interaction
contrasts are computed: Contrasts are generated for each factor separately,
one at a time; and these contrasts are applied to the object (the first time
around) or to the previous result (subsequently). (Any factors specified in
\code{by} are skipped.) The final result comprises contrasts of contrasts,
or, equivalently, products of contrasts for the factors involved. Processing
is done in the order of appearance in \code{object@levels}. With
\code{interaction = TRUE}, \code{method} (if specified as character) is used
for each contrast. If \code{interaction} is a character vector, the elements
specify the respective contrast method(s); they are recycled as needed.
}

\examples{
warp.lm <- lm(breaks ~ wool*tension, data = warpbreaks)
warp.emm <- emmeans(warp.lm, ~ tension | wool)
contrast(warp.emm, "poly")    # inherits 'by = "wool"' from warp.emm
pairs(warp.emm)               # ditto
contrast(warp.emm, "eff", by = NULL)  # contrasts of the 6 factor combs

# An interaction contrast for tension:wool
tw.emm <- contrast(warp.emm, interaction = c("poly", "consec"), by = NULL)
tw.emm          # see the estimates
coef(tw.emm)    # see the contrast coefficients
}
