% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff-size.R
\name{eff_size}
\alias{eff_size}
\title{Calculate effect sizes and confidence bounds thereof}
\usage{
eff_size(object, SD, SE.logSD, SE.SD, SE.var, LM, method = "pairwise",
  ...)
}
\arguments{
\item{object}{an \code{\link[=emmGrid-class]{emmGrid}} object, typically one defining the EMMs to 
be contrasted.}

\item{SD}{numeric scalar, value of the population SD. 
This may be missing only if \code{LM} is specified.}

\item{SE.logSD, SE.SD, SE.var}{numeric scalars that specify the standard
deviation of the sampling distribution of various functions of the
\code{SD} estimate -- respectively, that of \code{log(SD)}, \code{SD}
itself, or the variance, \code{SD^2}. The first non-missing value, in that
order, is the one that is used. All three of these may be missing if
\code{LM} is specified.}

\item{LM}{a model object having \code{\link{sigma}} and
\code{\link{df.residual}} methods. See Details for important considerations.}

\item{method}{the contrast method to use to define the effects.
This is passed to \code{\link{contrast}} after the elements of \code{object}
are scaled.}
}
\value{
an \code{\link[=emmGrid-class]{emmGrid}} object containing the effect sizes
}
\description{
Standardized effect sizes are typically calculated using pairwise differences of estimates,
divided by the SD of the population providing the context for those effects.
This function calculates effect sizes from an \code{emmGrid} object,
and confidence intervals for them, accounting for uncertainty in both the estimated
effects and the population SD. There is some flexibility in specifying the
needed parameters.
}
\details{
Any \code{by} variables specified in \code{object} will remain in force in the returned
effects, unless overridden in the optional arguments.

The \code{LM} parameter is recommended \emph{only} for models having a single
random effect, such as those fitted using \code{\link{lm}}. For more complex
models such as mixed models, the user usually should specify \code{SD} and
one of the \code{SE} parameters manually, based on some combination of the
random-effect variances.

It may be difficult to provide one of the \code{SE} parameters. But here are two 
useful things to know: (1) If \code{SD^2} is proportional to a chi-square random variable,
then \code{SE.logSD} equals \code{1 / sqrt(2 * df)}.
(2) \code{200 * SE.logSD} is roughly equal to the percentage range for most variations
in \code{SD}; for example, if you think \code{SD} is accurate to within 20 percent
of its true value, then \code{SE.logSD = 0.1} is about right (and from the first point, 
this corresponds to 50 degrees of freedom).

Specifying one of the \code{SE} parameters equal to zero is allowable, in which case you
are assuming that \code{SD} is known exactly. Obviously, this narrows the confidence
intervals for the effect sizes -- unrealistically if in fact \code{SD} is estimated.
}
\note{
The effects are always computed on the scale of the \emph{linear-predictor};
any response transformation or link function is completely ignored. If you
wish to base the effect sizes on the response scale, it is \emph{not} enough
to replace \code{object} with \code{regrid(object)}, because this
back-transformation changes the SD required to compute effect sizes. This is
a can of worms not worth opening, in my opinion.
}
\section{Computation}{

This function uses calls to \code{\link{regrid}} to put the estimated
marginal means (EMMs) on the log scale. Then an extra element is added to
this grid for the log of \code{SD} and is standard error (where we assume
that \code{SD} is uncorrelated with the log EMMs). Then a call to
\code{\link{contrast}} subtracts \code{log{SD}} from each of the log EMMs,
yielding values of \code{log(EMM/SD)}.
Finally, the results are re-gridded back to the original scale and the
desired contrasts are computed using \code{method}. In the log-scaling
part, we actually rescale the absolute values and keep track of the signs.
}

\examples{
fiber.lm <- lm(strength ~ diameter + machine, data = fiber)
emm <- emmeans(fiber.lm, "machine")
eff_size(emm, LM = fiber.lm)

# or equivalently:
eff_size(pairs(emm), LM = fiber.lm, method = "identity")

# compare with naively scaling by error SD without accounting for its uncertainty:
confint(pairs(emm, scale = 1/sigma(fiber.lm), adjust = "none"))
# This is more efficient than eff_size(emm, LM = fiber.lm, SE.logSD = 0)

# These results illustrate a sobering message that effect sizes are often
# not nearly as accurate as you may think.
}
