% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmGrid-methods.R
\name{str.emmGrid}
\alias{str.emmGrid}
\alias{print.emmGrid}
\alias{vcov.emmGrid}
\title{Miscellaneous methods for \code{emmGrid} objects}
\usage{
\method{str}{emmGrid}(object, ...)

\method{print}{emmGrid}(x, ...)

\method{vcov}{emmGrid}(object, ...)
}
\arguments{
\item{object}{An \code{emmGrid} object}

\item{...}{(required but not used)}

\item{x}{An \code{emmGrid} object}
}
\value{
The \code{vcov} method returns a symmetric matrix of variances and
  covariances for \code{predict.emmGrid(object, type = "lp")}
}
\description{
Miscellaneous methods for \code{emmGrid} objects
}
