% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmip.R
\name{emmip}
\alias{emmip}
\alias{emmip.default}
\title{Interaction-style plots for estimated marginal means}
\usage{
emmip(object, formula, ...)

\method{emmip}{default}(object, formula, type, pch = c(1, 2, 6, 7, 9, 10,
  15:20), lty = 1, col = NULL, plotit = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{emm}, or a fitted model of a class
supported by the \pkg{emmeans} package}

\item{formula}{Formula of the form 
\code{trace.factors ~ x.factors | by.factors}. The EMMs are
plotted against \code{x.factor} for each level of \code{trace.factors}.
\code{by.factors} is optional, but if present, it determines separate
panels. Each element of this formula may be a single factor in the model,
or a combination of factors using the \code{*} operator.}

\item{...}{Additional arguments passed to \code{\link{emmeans}} (when
\code{object} is not already an \code{emm} object), or to
\code{\link[lattice]{xyplot}}.}

\item{type}{As in \code{\link{predict.emm}}, this determines
whether we want to inverse-transform the predictions
(\code{type = "response"}) or not (any other choice). The default is
\code{"link"}, unless the \code{"predict.type"} option is in force; see
\code{\link{emm_options}}.}

\item{pch}{The plotting characters to use for each group (i.e., levels of
\code{trace.factors}). They are recycled as needed.}

\item{lty}{The line types to use for each group. Recycled as needed.}

\item{col}{The colors to use for each group, recycled as needed. If not
specified, the default trellis colors are used.}

\item{plotit}{Logical value. If \code{TRUE}, the plot is displayed.
Otherwise, one may use the \code{"lattice"} attribute of the returned
object and print it, perhaps after additional manipulation.}
}
\value{
(Invisibly), a \code{\link{data.frame}} with the table of
  least-squares means that were plotted, with an additional \code{"lattice"}
  attribute containing the \code{trellis} object for the plot.
}
\description{
Creates an interaction plot of EMMs based on a fitted model and a simple
formula specification.
}
\note{
This function uses the \code{\link[lattice]{xyplot}} function in the
  \code{lattice} package (an error is returned if \code{lattice} is not
  installed). Conceptually, it is equivalent to
  \code{\link{interaction.plot}} where the summarization function is thought
  to return the
  EMMs.
}
\section{Details}{

If \code{object} is a fitted model, \code{\link{emmeans}} is called with an
appropriate specification to obtain estimated marginal means for each
combination of the factors present in \code{formula} (in addition, any 
arguments in \code{\dots} that match \code{at}, \code{trend}, 
\code{cov.reduce}, or \code{fac.reduce} are passed to \code{emmeans}). 
Otherwise, if \code{object} is an \code{emm} object, its first element is 
used, and it must contain one estimate for each combination of the factors
present in \code{formula}.
}

\examples{
#--- Two-factor example
warp.lm <- lm(breaks ~ wool * tension, data = warpbreaks)

# Following plot is the same as the usual interaction plot of the data
emmip(warp.lm, wool ~ tension)

#--- Three-factor example
noise.lm = lm(noise ~ size * type * side, data = auto.noise)

# Separate interaction plots of size by type, for each side
emmip(noise.lm, type ~ size | side)

# One interaction plot, using combinations of size and side as the x factor
emmip(noise.lm, type ~ side * size)

# One interaction plot using combinations of type and side as the trace factor
# customize the colors, line types, and symbols to suggest these combinations
emmip(noise.lm, type * side ~ size, lty=1:2, col=1:2, pch=c(1,1,2,2))

# 3-way interaction is significant, but doesn't much visual difference...
noise.lm2 = update(noise.lm, . ~ . - size:type:side)
emmip(noise.lm2, type * side ~ size, lty=1:2, col=1:2, pch=c(1,1,2,2))

}
\seealso{
\code{\link{emmeans}}, \code{\link{interaction.plot}}
}
