% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{joint_tests}
\alias{joint_tests}
\title{Compute \dQuote{type III} tests of the terms in a model}
\usage{
joint_tests(object, by = NULL, ...)
}
\arguments{
\item{object}{a fitted model or an \code{emmGrid}. If a fitted model, it is
replaced by \code{ref_grid(object, cov.reduce = range, ...)}}

\item{by}{character names of \code{by} variables. Separate sets of tests are
run for each combination of these.}

\item{...}{additional arguments passed to \code{ref_grid} and \code{emmeans}}
}
\value{
a \code{summary_emm} object (same as is produced by 
  \code{\link{summary.emmGrid}})
}
\description{
This function produces an analysis-of-variance table based on linear
functions of predictors in a model or \code{emmGrid} object. Optionally, one
or more of the predictors may be used as a \dQuote{by} variable, so that
separate tables of tests are produced for each combination of them. When
equal weighting is used, the tests correspond to \dQuote{type III} tests such
as \pkg{SAS} produces; but other weighting schemes provide generalizations.
}
\details{
The \code{joint_tests} function constructs, for each combination of factors,
a set of (interaction) contrasts via \code{\link{contrast}}, and then tests
them using \code{\link{test}} with \code{joint = TRUE}. This is more robust
than the model-reduction approach used in \code{\link[car]{Anova}} in the
\pkg{car} package. In the latter, the user must fit the model using
sum-to-zero contrast specifications such as \code{contr.sum} in order to
obtain correct type-III tests. But \code{joint_tests} does them correctly
regardless of which contrast functions were used to construct the model
matrix.
}
\examples{
pigs.lm <- lm(log(conc) ~ source * factor(percent), data = pigs)

joint_tests(pigs.lm)                     ## type III ANOVA

joint_tests(pigs.lm, weights = "outer")  ## differently weighted

joint_tests(pigs.lm, by = "source")      ## separate type III tests of 'percent'
}
\seealso{
\code{\link{test}}
}
