% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwpp.R
\name{pwpp}
\alias{pwpp}
\title{Pairwise P-value plot}
\usage{
pwpp(emm, method = "pairwise", by, values = TRUE, rows = ".", xlab,
  ylab, xsub = "", ...)
}
\arguments{
\item{emm}{An \code{emmGrid} object}

\item{method}{Character or list. Passed to \code{\link{contrast}}, and defines 
the contrasts to be displayed. Any contrast method may be used,
provided that each contrast includes one coefficient of \code{1},
one coefficient of \code{-1}, and the rest \code{0}. That is, calling
\code{contrast(object, method)} produces a set of comparisons, each with
one estimate minus another estimate.}

\item{by}{Character vector of variable(s) in the grid to condition on. These will
create different panels, one for each level or level-combination.
Grid factors not in \code{by} are the \emph{primary} factors: 
whose levels or level combinations are compared pairwise.}

\item{values}{Logical value. If \code{TRUE}, the values of the EMMs are included
in the plot. When there are several side-by-side panels due
to \code{by} variable(s), the labels showing values start
stealing a lot of space from the plotting area; in those cases,
it may be desirable to specify \code{FALSE} or use \code{rows}
so that some panels are vertically stacked.}

\item{rows}{Character vector of which \code{by} variable(s) are used to define
rows of the panel layout. Those variables in \code{by} not included in 
\code{rows} define columns in the array of panels.
A \code{"."} indicates that only one row
is used, so all panels are stacked side-by-side.}

\item{xlab}{Character label to use in place of the default for the P-value axis.}

\item{ylab}{Character label to use in place of the default for the primary-factor axis.}

\item{...}{Additional arguments passed to \code{contrast} and \code{\link{summary.emmGrid}}}
}
\description{
Constructs a plot of P values associated with pairwise comparisons of 
estimated marginal means.
}
\details{
Factor levels (or combinations thereof) are plotted on the vertical scale, and P values
are plotted on the horizontal scale. Each P value is plotted twice -- at
vertical positions corresponding to the levels being compared -- and connected by
a line segment. Thus, it is easy to visualize which P values are small and large,
and which levels are compared. In addition, factor levels are color-coded, and the points
and half-line segments appear in the color of the other level.
The P-value scale is nonlinear, so as to stretch-out smaller P values and
compress larger ones.
}
\examples{
pigs.lm <- lm(log(conc) ~ source * factor(percent), data = pigs)
emm = emmeans(pigs.lm, ~ percent | source)
pwpp(emm)
pwpp(emm, method = "trt.vs.ctrl1", type = "response", side = ">")
}
