% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairpp.R
\name{pwpp}
\alias{pwpp}
\title{Pairwise p-value plot}
\usage{
pwpp(object, sort = TRUE, values = TRUE, by, rows = ".",
  adjust = "Tukey", xlab, ylab, ...)
}
\arguments{
\item{object}{An \code{emmGrid} object}

\item{sort}{Logical flag to determine whether means are sorted first}

\item{values}{Logical value. If \code{TRUE}, the values of the EMMs are included
in the plot. When there are several side-by-side panels due
to \code{by} variable(s), the labels showing values start
stealing a lot of space from the plotting area; in those cases,
it may be desiravle to specify \code{FALSE} or use \code{rows}
so that some panels are vertically stacked.}

\item{by}{Character vector of variable(s) in the grid to condition on. These will
create different panels, one for each level or level-combination.
Grid factors not in \code{by} are the \emph{primary} factors: 
whose levels or level combinations are compared pairwise.}

\item{rows}{Character vector of which \code{by} variable(s) are used to define
rows of the panel layout. A \code{"."} indicates that only one row
is used, so all panels are stacked side-by-side.}

\item{adjust}{Character \code{adjust} argument passed to \code{summary.emmGrid}.}

\item{xlab}{Character label to use in place of the default for the P-value axis.}

\item{ylab}{Character label to use in place of the default for the primary-factor axis.}

\item{...}{Additional arguments passed to \code{\link{summary.emmGrid}}}
}
\description{
Constructs a plot of P values associated with pairwise comparisons of 
estimated marginal means.
}
